changes to all files

+++ iscript/valc-install-part-2.sh
@@ -1,23 +1,26 @@
 #!/bin/bash
 
+
 exe () {
-    exit_code=1
-    while [ $exit_code -ne 0 ]; do
+    # exe hepls "contain a block of code in a repeatable format in case something goes wrong
+    return_code=1
+    while [ $return_code -ne 0 ]; do
 
-        $1
-
+        "$@"
+        
+        return_code=$?
         # check if everything worked
-        if [ $? -eq 0 ]; then
-            exit_code=0
+        if [ $return_code -eq 0 ]; then
             echo "Checks out"; sleep 1
             break
         fi
 
-        while true; do
+        while [ return_code -ne 0 ]; do
+            echo "Error code: $return_code !"
             read -p "Something went wrong here :( Do you want to redo the command? [y/n]: " yn
             case $yn in
                 [Yy]* ) break;;
-                [Nn]* ) exit_code=0; break;;
+                [Nn]* ) return_code=0; break;;
                 * ) echo "Enter 'y' or 'n'!";;
             esac
         done
@@ -28,15 +31,15 @@ notification () {
     clear; echo "$1"; sleep 1
 }
 
+
 ena_parallel () {
 
     # set a new index for parallel downloads
     # first install sed
-    notification "Installing sed"
-    pacman -S --noconfirm sed
-
-    notification "Enable parallel downloads"
+    notification "Installing sed and enabling parallel downloads"
+    pacman -S --noconfirm sed &&
     sed -i "s/^#ParallelDownloads = 5$/ParallelDownloads = 15/" /etc/pacman.conf 
+    [ $? -ne 0 ] && return 22 || : 
 }
 
 
@@ -44,17 +47,18 @@ time_setup () {
     
     # time
     notification "Time setup"
-    ln -sf /usr/share/zoneinfo/Europe/Berlin /etc/localtime
-
+    ln -sf /usr/share/zoneinfo/Europe/Berlin /etc/localtime &&
     hwclock --systohc
+    [ $? -ne 0 ] && return 23 || :
 }
 
 language_setup () {
     # language
     notification "Language setup"
-    echo "en_US.UTF-8 UTF-8" >> /etc/locale.gen
-    locale-gen
+    echo "en_US.UTF-8 UTF-8" >> /etc/locale.gen &&
+    locale-gen &&
     echo "LANG=en_US.UTF-8" > /etc/locale.conf
+    [ $? -ne 0 ] && return 24 || :
 }
 
 kb_setup () {
@@ -68,7 +72,11 @@ kb_setup () {
 
         case $yn in
 
-            [yY]* ) echo "KEYMAP=de-latin1" > /etc/vconsole.conf; break;;
+            [yY]* ) 
+                echo "KEYMAP=de-latin1" > /etc/vconsole.conf
+                [ $? -ne 0 ] && return 25 || :
+
+                break;;
             [nN]* ) echo "Not accessible yet"; sleep 2;;
             * ) echo "Enter 'y' or 'n'!";;
         esac
@@ -80,12 +88,30 @@ host_setup () {
     # Hostname
     notification "Hostname"
     echo "HOSTNAME: "
-    read hostname
-    echo $hostname > /etc/hostname
+    
+    read hostname 
 
-    echo "127.0.0.1       localhost" >> /etc/hosts
-    echo "::1             localhost" >> /etc/hosts
+    while true; do
+
+        read -p "Is it spelled correctly? [y/n]: " yn
+
+        case $yn in
+
+            [yY]* ) break;; 
+
+            [nN]* ) return 26;;
+
+            * ) echo "Enter 'y' or 'n'!";;
+        esac
+    done
+
+    echo $hostname > /etc/hostname &&
+
+    echo "127.0.0.1       localhost" >> /etc/hosts &&
+    echo "::1             localhost" >> /etc/hosts &&
     echo "127.0.1.1       $hostname.localdomain $hostname" >> /etc/hosts
+
+    [ $? -ne 0 ] && return 27 || :
 
 }
 
@@ -94,14 +120,39 @@ user_setup () {
     # User
     notification "User"
     passwd
+    [ $? -ne 0 ] && return 28 || :
+
     echo "USER: "
     read user
-    useradd -m $user
-    echo "Password for $user: "
+
+
+    while true; do
+
+        read -p "Is it spelled correctly? [y/n]: " yn
+
+        case $yn in
+
+            [yY]* ) break;; 
+
+            [nN]* ) return 29;;
+
+            * ) echo "Enter 'y' or 'n'!";;
+        esac
+    done
+
+    useradd -m $user 
+    [ $? -ne 0 ] && return 30 || :
+
+    echo "Password for $user: " &&
     passwd $user
+    [ $? -ne 0 ] && return 31 || :
+ 
     usermod -aG wheel,audio,video $user
-    pacman --noconfirm -S sudo
+    [ $? -ne 0 ] && return 32 || :
+
+    pacman --noconfirm -S sudo &&
     sudo sed -i '/^# %wheel ALL=(ALL:ALL) ALL$/s/^# //' /etc/sudoers | sudo EDITOR='tee -a' visudo
+    [ $? -ne 0 ] && return 33 || :
 
 }
 
@@ -109,6 +160,7 @@ inst_important () {
 
     notification "Installing important packages"
     pacman -S --noconfirm grub efibootmgr dosfstools os-prober mtools
+    [ $? -ne 0 ] && return 34 || :
 
 }
 
@@ -116,10 +168,14 @@ grub_setup () {
 
     # grub
     notification "Setting up grub"
-    grub-install --target=x86_64-efi --efi-directory=/boot/EFI --bootloader=GRUB --removable
+    grub-install --target=x86_64-efi --efi-directory=/boot/EFI --bootloader=GRUB --removable 
+    [ $? -ne 0 ] && return 35 || :
+
+
     # sed -i 's/quiet/pci=noaer/g' /etc/default/grub
     # sed -i 's/GRUB_TIMEOUT=5/GRUB_TIMEOUT=0/g' /etc/default/grub
     grub-mkconfig -o /boot/grub/grub.cfg
+    [ $? -ne 0 ] && return 36 || :
 
 }
 
@@ -127,17 +183,21 @@ network_manager () {
 
     # network manager 
     notification "Installing and setting up Networkmanager"
-    pacman --noconfirm -S networkmanager
+    pacman --noconfirm -S networkmanager 
+    [ $? -ne 0 ] && return 37 || :
+
     systemctl enable NetworkManager
+    [ $? -ne 0 ] && return 38 || :
 
 }
 
 inst_part () {
    
     notification "Installing next valc-installation-part"
-    curl https://raw.githubusercontent.com/d3ltaaa/valc/main/iscript/valc-install-part-3.sh > /home/$user/valc-install-part-3.sh
+    curl https://raw.githubusercontent.com/d3ltaaa/valc/main/iscript/valc-install-part-3.sh > /home/$user/valc-install-part-3.sh &&
     chmod +x /home/$user/valc-install-part-3.sh
-
+    [ $? -ne 0 ] && return 39 || :
+    
 }
 
 exe ena_parallel
